+++ src/app/pages/facsimile-collection-upload-selection/facsimile-collection-upload-selection.component.ts
@@ -287,7 +287,7 @@ export class FacsimileCollectionUploadSelectionComponent implements OnInit {
   }
 
   returnNav(facsCollId?: number | null): void {
-    const routeSegments = facsCollId ? ['/facsimiles', facsCollId] : ['/facsimiles'];
+    const routeSegments = facsCollId ? ['/facsimile-collections', facsCollId] : ['/facsimile-collections'];
     this.router.navigate(
       routeSegments,
       { queryParamsHandling: 'preserve' }